requirements.txt
Django==5.2.5
django-widget-tweaks==1.5.0
openpyxl==3.1.2
python-telegram-bot==20.6
python-dotenv==1.0.0
